moved email rule into separate file
