+++ backend/buildeR.Common/FluentValidators/User/NewUserDtoValidator.cs
@@ -1,4 +1,5 @@
 using buildeR.Common.DTO.User;
+using buildeR.Common.FluentValidators.Shared;
 using FluentValidation;
 
 namespace buildeR.Common.FluentValidators.User
@@ -10,7 +11,7 @@ namespace buildeR.Common.FluentValidators.User
             RuleFor(u => u.Username).Username();
             RuleFor(u => u.LastName).LastName();
             RuleFor(u => u.FirstName).FirstName();
-            RuleFor(u => u.Email).CustomEmail();
+            RuleFor(u => u.Email).Email();
             RuleFor(u => u.Bio).Bio();
         }
     }
