+++ backend/buildeR.Common/FluentValidators/User/UserDtoValidator.cs
@@ -1,6 +1,6 @@
 using buildeR.Common.DTO.User;
+using buildeR.Common.FluentValidators.Shared;
 using FluentValidation;
-using FluentValidation.Validators;
 
 namespace buildeR.Common.FluentValidators.User
 {
@@ -11,7 +11,7 @@ namespace buildeR.Common.FluentValidators.User
             RuleFor(u => u.Username).Username();
             RuleFor(u => u.LastName).LastName();
             RuleFor(u => u.FirstName).FirstName();
-            RuleFor(u => u.Email).CustomEmail();
+            RuleFor(u => u.Email).Email();
             RuleFor(u => u.Bio).Bio();
         }
     }
